+++ src/com/company/Engine/Player/Player.java
@@ -7,19 +7,19 @@ import static java.lang.System.out;
 
 public class Player {
 
-    private byte lvl;
-    private byte hp;
-    private byte mp;
+    private short lvl;
+    private short hp;
+    private short mp;
 
-    private byte str;
-    private byte agi;
-    private byte ene;
-    private byte vit;
+    private short str;
+    private short agi;
+    private short ene;
+    private short vit;
 
-    private byte attackMin;
-    private byte attackMax;
+    private short attackMin;
+    private short attackMax;
 
-    private byte def;
+    private short def;
     private String name;
     //Player selected prof
     private String prof;
@@ -34,8 +34,8 @@ public class Player {
 
         if (this.prof.equals(String.valueOf(warr))) {
             //DK attack set
-            this.attackMin = (byte) (this.str / 6);
-            this.attackMax = (byte) (this.str / 4);
+            this.attackMin = (short) (this.str / 6);
+            this.attackMax = (short) (this.str / 4);
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
@@ -45,9 +45,9 @@ public class Player {
 
         //SM defense set
         if (this.prof.equals(String.valueOf(mage))) {
-            this.def = (byte) (this.agi / 4);
+            this.def = (short) (this.agi / 4);
         }else if(this.prof.equals(String.valueOf(warr))) {
-            this.def = (byte) (this.agi / 3);
+            this.def = (short) (this.agi / 3);
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
@@ -56,9 +56,9 @@ public class Player {
 
         //ARCH Defense and Attack calc
         if(this.prof.equals(String.valueOf(arch))){
-            this.def = (byte) (this.agi / 10);
-            this.attackMin = (byte) ((this.str / 14) + (this.agi / 7));
-            this.attackMax = (byte) ((this.str / 8) + (this.agi / 4));
+            this.def = (short) (this.agi / 10);
+            this.attackMin = (short) ((this.str / 14) + (this.agi / 7));
+            this.attackMax = (short) ((this.str / 8) + (this.agi / 4));
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
@@ -69,15 +69,15 @@ public class Player {
         //SM vit set
         //HP = 30+(lvl-1)+vit*2
         if (this.prof.equals(String.valueOf(mage))) {
-            this.hp = (byte) (30 + (this.lvl - 1) + (this.vit * 2));
+            this.hp = (short) (30 + (this.lvl - 1) + (this.vit * 2));
         }else if (this.prof.equals(String.valueOf(warr))) {
             //DK vit set
             //HP = 35+(lvl-1)*2+vit*3
-            this.hp = (byte) (35 + (this.lvl - 1) * 2 + (this.vit * 3));
+            this.hp = (short) (35 + (this.lvl - 1) * 2 + (this.vit * 3));
         }else if (this.prof.equals(String.valueOf(arch))) {
             //ARCH vit set
             //HP = 40+(lvl-1)+vit*2
-            this.hp = (byte) (40 + (this.lvl - 1) + this.vit * 2);
+            this.hp = (short) (40 + (this.lvl - 1) + this.vit * 2);
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
@@ -87,11 +87,11 @@ public class Player {
              //DK ene set
         if (this.prof.equals(String.valueOf(warr))) {
 
-            this.mp = (byte) (10 + (this.lvl - 1) * 0.5 + this.ene);
+            this.mp = (short) (10 + (this.lvl - 1) * 0.5 + this.ene);
 
         }else if (this.prof.equals(String.valueOf(arch))) {
             //ARCH ene set
-            this.mp = (byte) (6 + (this.lvl * 1.5) + (this.ene * 1.5));
+            this.mp = (short) (6 + (this.lvl * 1.5) + (this.ene * 1.5));
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
@@ -101,110 +101,110 @@ public class Player {
 
         //MAGE Mana and MagicAttack calc
         if (this.prof.equals(String.valueOf(mage))) {
-            this.attackMin = (byte) (this.ene / 9);
-            this.attackMax = (byte) (this.ene / 4);
-            this.mp = (byte) ((this.lvl - 1) * 2 + (this.ene * 2));
+            this.attackMin = (short) (this.ene / 9);
+            this.attackMax = (short) (this.ene / 4);
+            this.mp = (short) ((this.lvl - 1) * 2 + (this.ene * 2));
         }else {
             throw new Error("Method: \n" + Thread.currentThread().getStackTrace()[1] + "is unavailable for: " + this.prof + "\n");
         }
     }
 
 
-    public byte getLvl() {
+    public short getLvl() {
         return lvl;
     }
 
-    public void setLvl(byte lvl) {
+    public void setLvl(short lvl) {
         this.lvl = lvl;
     }
 
 
 
-    public byte getHp() {
+    public short getHp() {
         return hp;
     }
 
-    public void setHp(byte hp) {
+    public void setHp(short hp) {
         this.hp = hp;
     }
 
 
 
-    public byte getMp() {
+    public short getMp() {
         return mp;
     }
 
-    public void setMp(byte mp) {
+    public void setMp(short mp) {
         this.mp = mp;
     }
 
 
 
-    public byte getStr() {
+    public short getStr() {
         return str;
     }
 
-    public void setStr(byte str) {
+    public void setStr(short str) {
         this.str = str;
     }
 
 
 
-    public byte getAgi() {
+    public short getAgi() {
         return agi;
     }
 
-    public void setAgi(byte agi) {
+    public void setAgi(short agi) {
         this.agi = agi;
     }
 
 
 
-    public byte getEne() {
+    public short getEne() {
         return ene;
     }
 
-    public void setEne(byte ene) {
+    public void setEne(short ene) {
         this.ene = ene;
     }
 
 
 
-    public byte getVit() {
+    public short getVit() {
         return vit;
     }
 
-    public void setVit(byte vit) {
+    public void setVit(short vit) {
         this.vit = vit;
     }
 
 
 
-    public byte getAttackMin() {
+    public short getAttackMin() {
         return attackMin;
     }
 
-    public void setAttackMin(byte attackMin) {
+    public void setAttackMin(short attackMin) {
         this.attackMin = attackMin;
     }
 
 
 
-    public byte getAttackMax() {
+    public short getAttackMax() {
         return attackMax;
     }
 
-    public void setAttackMax(byte attackMax) {
+    public void setAttackMax(short attackMax) {
         this.attackMax = attackMax;
     }
 
 
 
-    public byte getDef() {
+    public short getDef() {
         return def;
     }
 
-    public void setDef(byte def) {
+    public void setDef(short def) {
         this.def = def;
     }
